desin updated, mathproblems updated,

+++ src/design/Employee.java
@@ -18,9 +18,12 @@ public interface Employee {
     public void assignDepartment();
 	
 	//calculate employee salary
-	public int calculateSalary();
+	public double calculateSalary();
 	
 	//employee benefit
 	public void benefitLayout();
+
+
+
 
 }
